Move ProcessInstance into dedicated file
