+++ include/Basil/Process.hpp
@@ -1,6 +1,7 @@
 #ifndef INCLUDE_BASIL_PROCESS_HPP_
 #define INCLUDE_BASIL_PROCESS_HPP_
 
+#include "Process/ProcessInstance.hpp"
 #include "Process/ProcessSchedule.hpp"
 #include "Process/TimeSource.hpp"
